+++ src/java/org/apache/cassandra/metrics/DecayingEstimatedHistogramReservoir.java
@@ -24,7 +24,6 @@ import java.io.PrintWriter;
 import java.nio.charset.Charset;
 import java.util.Arrays;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicLongArray;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
@@ -518,19 +517,23 @@ public class DecayingEstimatedHistogramReservoir implements Reservoir
 
             final long count = count();
 
-            if(count <= 1) {
+            if(count <= 1)
+            {
                 return 0.0D;
-            } else {
+            }
+            else
+            {
                 double mean = this.getMean();
                 double sum = 0.0D;
 
-                for(int i = 0; i < lastBucket; ++i) {
+                for(int i = 0; i < lastBucket; ++i)
+                {
                     long value = bucketOffsets[i];
-                    double diff = (double)value - mean;
+                    double diff = value - mean;
                     sum += diff * diff * decayingBuckets[i];
                 }
 
-                return Math.sqrt(sum / (double)(count - 1));
+                return Math.sqrt(sum / (count - 1));
             }
         }
 
@@ -540,7 +543,8 @@ public class DecayingEstimatedHistogramReservoir implements Reservoir
             {
                 int length = decayingBuckets.length;
 
-                for(int i = 0; i < length; ++i) {
+                for(int i = 0; i < length; ++i)
+                {
                     out.printf("%d%n", decayingBuckets[i]);
                 }
             }
